+++ packages/nimble-components/src/table/template.ts
@@ -1,6 +1,6 @@
 import {
     children,
-    ElementsFilter,
+    elements,
     html,
     ref,
     repeat,
@@ -11,30 +11,20 @@ import type { VirtualItem } from '@tanstack/virtual-core';
 import type { Table } from '.';
 import { TableHeader } from './components/header';
 import { TableRow } from './components/row';
-import { TableColumn } from '../table-column/base';
-
-const isTableColumn = (): ElementsFilter => {
-    const filter: ElementsFilter = (
-        value: Node,
-        _: number,
-        __: Node[]
-    ): boolean => {
-        return value instanceof TableColumn;
-    };
-    return filter;
-};
+import type { TableColumn } from '../table-column/base';
 
 // prettier-ignore
 export const template = html<Table>`
-    <template role="table" ${children({ property: 'columns', filter: isTableColumn() })}>
+    <template role="table" ${children({ property: 'childItems', filter: elements() })}>
         <div class="table-container">
-            <div role="rowgroup" class="header-container" style="margin-right: ${x => x.virtualizer.headerContainerMarginRight}px;">
+            <div role="rowgroup" class="header-container">
                 <div class="header-row" role="row">
                     ${repeat(x => x.columns, html<TableColumn>`
                         <${DesignSystem.tagFor(TableHeader)} class="header">
-                            ${x => x.textContent}
+                            <slot name="${x => x.slot}"></slot>
                         </${DesignSystem.tagFor(TableHeader)}>
                     `)}
+                    <div class="header-scrollbar-spacer" style="width: ${x => x.virtualizer.headerContainerMarginRight}px;"></div>
                 </div>
             </div>
             <div class="table-viewport" ${ref('viewport')}>