+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Fri, 10 Feb 2023 15:26:34 GMT and should not be manually modified.
+This log was last generated on Tue, 14 Feb 2023 15:44:48 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 16.0.14
+
+Tue, 14 Feb 2023 15:44:48 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v18.3.3
+
+## 16.0.13
+
+Mon, 13 Feb 2023 23:05:12 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v18.3.2
+
+## 16.0.12
+
+Fri, 10 Feb 2023 20:11:53 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v18.3.1
+
+## 16.0.11
+
+Fri, 10 Feb 2023 19:40:36 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v18.3.0
 
 ## 16.0.10